+++ backend/src/validation/orders.ts
@@ -13,7 +13,7 @@ import { z } from 'zod';
 export const orderItemSchema = z.object({
     productId: z.string().uuid('ID do produto inválido'),
     productName: z.string().max(200, 'Nome do produto muito longo').optional(),
-    quantity: z.coerce.number().int().positive('Quantidade deve ser maior que zero'),
+    quantity: z.coerce.number().positive('Quantidade deve ser maior que zero'),
     unitPrice: z.coerce.number().positive('Preço unitrio deve ser maior que zero').optional(),
     price: z.coerce.number().positive('Preço deve ser maior que zero').optional(),
     notes: z.string().max(200, 'Notas muito longas').optional().nullable()
